+++ src/fknni/__init__.py
@@ -2,4 +2,6 @@ from importlib.metadata import version
 
 __all__ = ["faiss"]
 
+from .faiss import FaissImputer
+
 __version__ = version("fknni")